feat: Add normalization and opponent simulation constants to enhance RL environment
